chore: remove v1alpha5 hash and disruption controllers
